source/mailing/requirements.txt
pandas
awswrangler
psycopg2
python-dotenv
boto3